+++ CartoMapAndroid/app/src/main/java/com/carto/cartomap/android/map/CartoUTFGridActivity.java
@@ -30,6 +30,7 @@ public class CartoUTFGridActivity extends VectorMapSampleBaseActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+
         // MapSampleBaseActivity creates and configures mapView  
         super.onCreate(savedInstanceState);
 
@@ -41,20 +42,23 @@ public class CartoUTFGridActivity extends VectorMapSampleBaseActivity {
             String statTag = "3c6f224a-c6ad-11e5-b17e-0e98b61680bf";
             String[] columns = new String[]{"name", "field_9", "slot"};
             String cartoCss =
-                    "#stations_1{marker-fill-opacity:0.9;marker-line-color:#FFF;marker-line-width:2;marker-line-opacity:1;marker-placement:point;marker-type:ellipse;marker-width:10;marker-allow-overlap:true;}\n" +
-                            "#stations_1[status = 'In Service']{marker-fill:#0F3B82;}\n" +
-                            "#stations_1[status = 'Not In Service']{marker-fill:#aaaaaa;}\n" +
-                            "#stations_1[field_9 = 200]{marker-width:80.0;}\n" +
-                            "#stations_1[field_9 <= 49]{marker-width:25.0;}\n" +
-                            "#stations_1[field_9 <= 38]{marker-width:22.8;}\n" +
-                            "#stations_1[field_9 <= 34]{marker-width:20.6;}\n" +
-                            "#stations_1[field_9 <= 29]{marker-width:18.3;}\n" +
-                            "#stations_1[field_9 <= 25]{marker-width:16.1;}\n" +
-                            "#stations_1[field_9 <= 20.5]{marker-width:13.9;}\n" +
-                            "#stations_1[field_9 <= 16]{marker-width:11.7;}\n" +
-                            "#stations_1[field_9 <= 12]{marker-width:9.4;}\n" +
-                            "#stations_1[field_9 <= 8]{marker-width:7.2;}\n" +
-                            "#stations_1[field_9 <= 4]{marker-width:5.0;}";
+                    "#stations_1{marker-fill-opacity:0.9;marker-line-color:#FFF;" +
+                            "marker-line-width:2;marker-line-opacity:1;marker-placement:point;" +
+                            "marker-type:ellipse;marker-width:10;marker-allow-overlap:true;" +
+                    "}\n" +
+                    "#stations_1[status = 'In Service']{marker-fill:#0F3B82;}\n" +
+                    "#stations_1[status = 'Not In Service']{marker-fill:#aaaaaa;}\n" +
+                    "#stations_1[field_9 = 200]{marker-width:80.0;}\n" +
+                    "#stations_1[field_9 <= 49]{marker-width:25.0;}\n" +
+                    "#stations_1[field_9 <= 38]{marker-width:22.8;}\n" +
+                    "#stations_1[field_9 <= 34]{marker-width:20.6;}\n" +
+                    "#stations_1[field_9 <= 29]{marker-width:18.3;}\n" +
+                    "#stations_1[field_9 <= 25]{marker-width:16.1;}\n" +
+                    "#stations_1[field_9 <= 20.5]{marker-width:13.9;}\n" +
+                    "#stations_1[field_9 <= 16]{marker-width:11.7;}\n" +
+                    "#stations_1[field_9 <= 12]{marker-width:9.4;}\n" +
+                    "#stations_1[field_9 <= 8]{marker-width:7.2;}\n" +
+                    "#stations_1[field_9 <= 4]{marker-width:5.0;}";
 
             // you probably do not need to change much of below
             configJson.put("version", "1.0.1");
@@ -74,6 +78,7 @@ public class CartoUTFGridActivity extends VectorMapSampleBaseActivity {
             JSONObject attributesJson = new JSONObject();
             attributesJson.put("id", "cartodb_id");
             JSONArray columnsJson = new JSONArray();
+
             for (String col : columns) {
                 columnsJson.put(col);
             }
@@ -83,41 +88,49 @@ public class CartoUTFGridActivity extends VectorMapSampleBaseActivity {
             layersJson.put("options", optionsJson);
             layersArrayJson.put(layersJson);
             configJson.put("layers", layersArrayJson);
+
         } catch (JSONException e) {
+
         }
 
 
         final String config = configJson.toString();
 
-        System.out.println(config);
         // Use the Maps service to configure layers. Note that this must be done
         // in a separate thread on Android, as Maps API requires connecting to server
         // which is not allowed in main thread.
 		Thread serviceThread = new Thread(new Runnable() {
 			@Override
 			public void run() {
-				CartoMapsService mapsService = new CartoMapsService();
+
+                CartoMapsService mapsService = new CartoMapsService();
 				mapsService.setUsername("nutiteq");
 				mapsService.setDefaultVectorLayerMode(true); // use vector layers
+
                 try {
                     LayerVector layers = mapsService.buildMap(Variant.fromString(config));
 
                     LocalVectorDataSource vectorDataSource = new LocalVectorDataSource(baseProjection);
                     VectorLayer vectorLayer = new VectorLayer(vectorDataSource);
-                    for (int i = 0; i < layers.size(); i++) {
-                        TileLayer layer = (TileLayer) layers.get(i);
-                        TileDataSource ds = layer.getUTFGridDataSource();
-                        MyUTFGridEventListener mapListener = new MyUTFGridEventListener(mapView, vectorDataSource);
-                        layer.setUTFGridEventListener(mapListener);
-                        mapView.getLayers().add(layer);
-                    }
+
                     mapView.getLayers().add(vectorLayer);
+
+                    for (int i = 0; i < layers.size(); i++) {
+
+                        TileLayer layer = (TileLayer) layers.get(i);
+                        mapView.getLayers().add(layer);
+
+                        MyUTFGridEventListener mapListener = new MyUTFGridEventListener(vectorLayer);
+                        layer.setUTFGridEventListener(mapListener);
+                    }
+
                 }
                 catch (IOException e) {
                     Log.e("EXCEPTION", "Exception: " + e);
                 }
 			}
 		});
+
 		serviceThread.start();
 
         // finally animate map to the content area
