+++ CartoMapAndroid/app/src/main/java/com/carto/cartomap/android/maplistener/MyUTFGridEventListener.java
@@ -3,8 +3,11 @@ package com.carto.cartomap.android.maplistener;
 import android.util.Log;
 
 import com.carto.core.MapPos;
+import com.carto.core.Variant;
 import com.carto.datasources.LocalVectorDataSource;
 import com.carto.layers.UTFGridEventListener;
+import com.carto.layers.VectorLayer;
+import com.carto.styles.BalloonPopupMargins;
 import com.carto.styles.BalloonPopupStyleBuilder;
 import com.carto.ui.ClickType;
 import com.carto.ui.MapView;
@@ -16,7 +19,8 @@ import java.util.Locale;
 /**
  * A custom map event listener that displays information about map events and creates pop-ups.
  */
-public class MyUTFGridEventListener extends UTFGridEventListener {
+
+/*public class MyUTFGridEventListener extends UTFGridEventListener {
 	private MapView mapView;
 	private LocalVectorDataSource vectorDataSource;
 	
@@ -30,7 +34,8 @@ public class MyUTFGridEventListener extends UTFGridEventListener {
 	@Override
 	public boolean onUTFGridClicked(UTFGridClickInfo clickInfo) {
 		Log.d("LOG", "UTF grid click!");
-		
+
+		System.out.println("onUTFGridClicked");
 		// Remove old click label
 		if (oldClickLabel != null) {
 			vectorDataSource.remove(oldClickLabel);
@@ -58,19 +63,52 @@ public class MyUTFGridEventListener extends UTFGridEventListener {
 
 		String msg = clickInfo.getElementInfo().toString();
 
-		// finally show click coordinates also
+		// Show click coordinates also
 		MapPos wgs84Clickpos = mapView.getOptions().getBaseProjection().toWgs84(clickPos);
 		msg  += String.format(Locale.US, "%.4f, %.4f", wgs84Clickpos.getY(), wgs84Clickpos.getX());
 
+		BalloonPopup clickPopup = new BalloonPopup(clickInfo.getClickPos(), styleBuilder.buildStyle(), clickMsg, msg);
 
-
-		BalloonPopup clickPopup = new BalloonPopup(clickInfo.getClickPos(),
-												   styleBuilder.buildStyle(),
-		                						   clickMsg,
-		                						   msg);
 		vectorDataSource.add(clickPopup);
 		oldClickLabel = clickPopup;
 		
 		return true;
 	}
 }
+*/
+
+public class MyUTFGridEventListener extends UTFGridEventListener {
+	private VectorLayer vectorLayer;
+
+	public MyUTFGridEventListener(VectorLayer vectorLayer) {
+		this.vectorLayer = vectorLayer;
+	}
+
+	@Override
+	public boolean onUTFGridClicked(UTFGridClickInfo utfGridClickInfo) {
+		LocalVectorDataSource vectorDataSource = (LocalVectorDataSource) vectorLayer.getDataSource();
+
+		System.out.println("UtfGridClicked");
+
+		// Clear previous popups
+		vectorDataSource.clear();
+
+		BalloonPopupStyleBuilder styleBuilder = new BalloonPopupStyleBuilder();
+
+		// Configure style
+		styleBuilder.setLeftMargins(new BalloonPopupMargins(0, 0, 0, 0));
+		styleBuilder.setTitleMargins(new BalloonPopupMargins(6, 3, 6, 3));
+
+		// Make sure this label is shown on top all other labels
+		styleBuilder.setPlacementPriority(10);
+
+		// Show clicked element variant as JSON string
+		String desc = utfGridClickInfo.getElementInfo().toString();
+
+		BalloonPopup clickPopup = new BalloonPopup(utfGridClickInfo.getClickPos(), styleBuilder.buildStyle(), "Clicked", desc);
+		vectorDataSource.add(clickPopup);
+
+		return true;
+	}
+}
+
